% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capacityNumbers}
\alias{capacityNumbers}
\title{Get max capacity for 1.0 and 1.9 ml tubes}
\usage{
capacityNumbers(x = readKBExcel())
}
\arguments{
\item{x}{input data from Karen Beeri about BioStore capacity and fullness}
}
\value{
a list of MAXIMUM possible tubes of each size that can go in the BioStore
}
\description{
Get max capacity for 1.0 and 1.9 ml tubes
}
\examples{
capacityNumbers()
}
