% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_full_date}
\alias{find_full_date}
\title{Function to find the date the freezer is full. Under development!}
\usage{
find_full_date(predictions, capacity, resultslm)
}
\arguments{
\item{predictions}{predictions}

\item{capacity}{capacity}

\item{resultslm}{model}
}
\value{
date when freezer will be full. function in development and doesn't do much yet
}
\description{
Function to find the date the freezer is full. Under development!
}
