% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freezer_fullness_graph}
\alias{freezer_fullness_graph}
\title{Make a very simple plot to visually represent how full and empty freezer is at a given time}
\usage{
freezer_fullness_graph(used = NULL)
}
\arguments{
\item{used}{Percent of freezer currently occupied}
}
\value{
A ggplot2 plot
}
\description{
Make a very simple plot to visually represent how full and empty freezer is at a given time
}
