% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_projections.R
\name{future_projections}
\alias{future_projections}
\title{Projection of future collections}
\usage{
future_projections(pt = participants_proj(), bc = readCollections())
}
\arguments{
\item{pt}{Participant projections}

\item{bc}{Biospecimen collection protocol}
}
\value{
a tibble
}
\description{
Projection of future collections
}
\examples{
future_projections()
}
