% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kits_proj}
\alias{kits_proj}
\title{Load in projected kit count numbers from PTF August 2025}
\usage{
kits_proj(x = "PTF_projections_Aug2025.csv", make_long = FALSE)
}
\arguments{
\item{x}{file name}

\item{make_long}{make returned tibble long instead of wide}
}
\value{
a tibble with this information
}
\description{
Load in projected kit count numbers from PTF August 2025
}
\examples{
kits_proj()
}
