% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longifySuchiBAM}
\alias{longifySuchiBAM}
\title{Make Suchi's historical data "long" instead of "wide"}
\usage{
longifySuchiBAM(x = readSuchiBAM())
}
\arguments{
\item{x}{"Wide" version of historical data of ECHO submissions to BioStore}
}
\value{
"Long" version of historical data of ECHO submissions to BioStore
}
\description{
Make Suchi's historical data "long" instead of "wide"
}
\examples{
longifySuchiBAM()
}
