% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_ptf_tube_counts}
\alias{merge_ptf_tube_counts}
\title{Merge PTF kit projection counts with number of tubes in each kit}
\usage{
merge_ptf_tube_counts(x = NULL, y = NULL)
}
\arguments{
\item{x}{PTF projected kit counts}

\item{y}{Number of tubes in each kit}
}
\value{
tibble merged of both above
}
\description{
Merge PTF kit projection counts with number of tubes in each kit
}
\examples{
merge_ptf_tube_counts()
}
