% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{participants_proj}
\alias{participants_proj}
\title{Load in predicted participant numbers from Measurement Core May 2025}
\usage{
participants_proj(x = "MC_projections_May2025.csv")
}
\arguments{
\item{x}{file name}
}
\value{
a tibble with this information
}
\description{
Load in predicted participant numbers from Measurement Core May 2025
}
\examples{
participants_proj()
}
