% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pendingNumbers}
\alias{pendingNumbers}
\title{Pending numbers}
\usage{
pendingNumbers(x = readKBExcel())
}
\arguments{
\item{x}{input data from Suchi and Karen Beeri about tubes pending storage}
}
\value{
a list of pending tubes of each size slated to go in the BioStore
}
\description{
Pending numbers
}
\examples{
pendingNumbers()
}
