% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCollections}
\alias{readCollections}
\title{Load in a file with current biospecimen collection protocol and timeline}
\usage{
readCollections(x = "biospecimen_collection_for_biostore_calculations.xlsx")
}
\arguments{
\item{x}{File name}
}
\value{
A tibble with biospecimen collection information
}
\description{
Load in a file with current biospecimen collection protocol and timeline
}
\examples{
readCollections()
}
