% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readcollectionsdescriptions}
\alias{readcollectionsdescriptions}
\title{Description of information in readCollections()}
\usage{
readcollectionsdescriptions()
}
\value{
a gt table
}
\description{
Description of information in readCollections()
}
\examples{
readcollectionsdescriptions()
}
