% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readKBExcel}
\alias{readKBExcel}
\title{Load in BioStore capacity Excel file from Karen Beeri}
\usage{
readKBExcel(x = "calculator_for_Suman.xlsx")
}
\arguments{
\item{x}{Exact name of the file}
}
\value{
A clean tibble of the Excel file for further manipulation
}
\description{
Load in BioStore capacity Excel file from Karen Beeri
}
\examples{
readKBExcel()
}
