% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSuchiBAM}
\alias{readSuchiBAM}
\title{Read in file from Suchi with historical data on ECHO submissions to the BioStore}
\usage{
readSuchiBAM(x = "suchi_bam_submissions.csv")
}
\arguments{
\item{x}{File name}
}
\value{
A tibble with historical data on ECHO submissions to the BioStore
}
\description{
Read in file from Suchi with historical data on ECHO submissions to the BioStore
}
\examples{
readSuchiBAM()
}
