% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_arima.R
\name{single_arima}
\alias{single_arima}
\title{Single ARIMA model with covariates and two scenarios (initial stock and final jump)}
\usage{
single_arima()
}
\description{
Single ARIMA model with covariates and two scenarios (initial stock and final jump)
}
