% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{site_collections}
\alias{site_collections}
\title{Load in de-identified information about site biospecimen collections. Only captures specimens NOT yet shipped to biorepository.}
\usage{
site_collections(x = NULL, biostore_only = TRUE)
}
\arguments{
\item{x}{file name of deidentified biospecimen collections file}

\item{biostore_only}{Pull only specimens collected that can go in the BioStore or not.}
}
\value{
a tibble with this information
}
\description{
Load in de-identified information about site biospecimen collections. Only captures specimens NOT yet shipped to biorepository.
}
\examples{
site_collections()
}
