% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteCollections.R
\name{siteCollections}
\alias{siteCollections}
\title{Parse file of site collections from Bio-Track}
\usage{
siteCollections(x = "site_report_2025_09_11_07_43_04.json")
}
\arguments{
\item{x}{JSON file from Bio-Track}
}
\description{
Parse file of site collections from Bio-Track
}
\examples{
siteCollections()
}
