% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteCollections.R
\name{siteCollectionsPre}
\alias{siteCollectionsPre}
\title{Parse file of site collections from Bio-Track}
\usage{
siteCollectionsPre(x = "site_report_2025_09_12_07_43_04.json")
}
\arguments{
\item{x}{JSON file from Bio-Track}
}
\value{
tibble with collections from sites that will go into biostore
}
\description{
Parse file of site collections from Bio-Track
}
