% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{totalBioStoreCapacity}
\alias{totalBioStoreCapacity}
\title{Calculate total BioStore capacity}
\usage{
totalBioStoreCapacity(x = NULL, y = NULL)
}
\arguments{
\item{x}{number of 1.0 ml tubes that could be added to storage}

\item{y}{number of 1.9 ml tubes that could be added to storage}
}
\value{
nothing yet
}
\description{
Calculate total BioStore capacity
}
\examples{
totalBioStoreCapacity(x = 100, y = 100) # function doesn't do anything yet
}
