% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zoo_ts}
\alias{zoo_ts}
\title{Create a zoo ts object from historical data
https://cran.r-project.org/web/packages/zoo/index.html}
\usage{
zoo_ts(x = readHistorical(), tube = "cumulative_1.0")
}
\arguments{
\item{x}{Historical data}

\item{tube}{tube size, and whether cumulative or not. Options are "tubes_1.0_ml", "tubes_1.9_ml", "cumulative_1.0" or "cumulative_1.9"}
}
\value{
zoo ts object
}
\description{
Create a zoo ts object from historical data
https://cran.r-project.org/web/packages/zoo/index.html
}
\examples{
zoo_ts()
}
