% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_arima.R
\name{predict_capacity}
\alias{predict_capacity}
\title{Code is all Eric Koplin!}
\usage{
predict_capacity(hist_init, horizon = 6)
}
\value{
arima
}
\description{
Code is all Eric Koplin!
}
\examples{
predict_capacity()
}
