% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{site_collections}
\alias{site_collections}
\title{Load in de-identified information about site biospecimen collections}
\usage{
site_collections(x = "deidentified_specimen_collection.csv")
}
\arguments{
\item{x}{file name of deidentified biospecimen collections file}
}
\value{
a tibble with this information
}
\description{
Load in de-identified information about site biospecimen collections
}
\examples{
site_collections()
}
